requirements.txt
langchain-huggingface
langchain
langchain-community
transformers
faiss-cpu
bitsandbytes
accelerate
sentence-transformers
datasets
pymupdf
kiwipiepy
pandas
torch
python-dotenv
numpy
tqdm
fitz
scikit-learn